User can now only delete their own listings

+++ controllers/ListingController.js
@@ -13,6 +13,7 @@ export const createListing = async (req, res) => {
         if (String(req.user.user_id) != String(userProperty.user_id)) {
             return res.status(403).json({ message: 'User does not own this property' });
         }
+
         const listingData = req.body;
 
         const listing = await ListingService.createListing(listingData);
@@ -50,6 +51,15 @@ export const updateListing = async (req, res) => {
     const updateData = req.body;
 
     try {
+
+        const userProperty = await UserProperty.findOne({
+            where: {property_id: req.body.property_id }
+        })
+
+        if (String(req.user.user_id) != String(userProperty.user_id)) {
+            return res.status(403).json({ message: 'User does not own this property' });
+        }
+
         const updatedListing = await ListingService.updateListing(listingId, updateData);
         if (!updatedListing) {
             return res.status(404).json({ message: 'Listing not found' });
@@ -68,6 +78,17 @@ export const deleteListing = async (req, res) => {
         if (!listing) {
             return res.status(404).json({ message: 'Listing not found' });
         }
+
+        const propertyId = listing.property_id;
+
+        const userProperty = await UserProperty.findOne({
+            where: { property_id: propertyId }
+        })
+
+        if (String(req.user.user_id) != String(userProperty.user_id)) {
+            return res.status(403).json({ message: 'User does not own this property' });
+        }
+
         await ListingService.deleteListing(listingId);
         return res.status(200).json({ message: `Listing with ID ${listingId} deleted successfully` });
     } catch (error) {
